GYCS-Charity-Vue: 更新了公益灾区活动和公益募资活动页面

+++ GYCS-Admin/gycs-admin/src/main/java/com/ruoyi/web/controller/charity/front/RaiseFundController.java
@@ -1,7 +1,6 @@
 package com.ruoyi.web.controller.charity.front;
 
 
-import com.alibaba.fastjson2.util.Fnv;
 import com.ruoyi.charity.domain.dto.CharityRaiseFund;
 import com.ruoyi.charity.domain.vo.BankTransferRecordVo;
 import com.ruoyi.charity.domain.vo.CertificateInfoVo;
@@ -82,7 +81,7 @@ public class RaiseFundController  extends BaseController {
      * @return AjaxResult
      */
     @GetMapping("/getCertificateInfo")
-    public AjaxResult getCertificateInfo(@RequestParam("raiseId") Long raiseId) {
+    public AjaxResult getCertificateInfo(@RequestParam("raiseId") BigInteger raiseId) {
         LoginUser loginUser = SecurityUtils.getLoginUser();
         return fundService.getCertificateInfo(raiseId,loginUser.getUsername());
     }
