+++ GYCS-Admin/gycs-charity/src/main/java/com/ruoyi/charity/service/RaiseFundService.java
@@ -17,7 +17,7 @@ public interface RaiseFundService {
 
     AjaxResult getRaiseFundInfo(Long raiseId,String username);
 
-    AjaxResult getCertificateInfo(Long raiseId,String username);
+    AjaxResult getCertificateInfo(BigInteger raiseId,String username);
 
     AjaxResult getRaiseFundVoteStatus(Long raiseId, String username);
